refactor: test 패키지의 support 패키지 정리
