+++ backend/src/test/java/com/woowacourse/kkogkkog/support/application/ServiceTest.java
@@ -1,6 +1,7 @@
-package com.woowacourse.kkogkkog.support.setup;
+package com.woowacourse.kkogkkog.support.application;
 
 import com.woowacourse.kkogkkog.infrastructure.application.SlackClient;
+import com.woowacourse.kkogkkog.support.common.DatabaseCleaner;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
